requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
contourpy==1.3.0
cycler==0.12.1
fonttools==4.55.0
idna==3.10
importlib_resources==6.4.5
joblib==1.4.2
kiwisolver==1.4.7
matplotlib==3.9.2
nltk==3.9.1
numpy
oauthlib==3.2.2
packaging==24.2
pillow==11.0.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==1.3.1
six==1.16.0
textblob==0.18.0.post0
tqdm==4.67.0
tweepy==4.14.0
urllib3==2.2.3
zipp==3.21.0
yfinance
